🚚build: PaddingLabel 컴포넌트도 분리
